requirements.txt
discord.py
git+https://github.com/Gorialis/jishaku
asyncpg
tabulate
psutil
requests  # Temporary due to SSL error with aiohttp
python-dotenv
